v1 API

+++ src/main/java/com/hipervet/api/repositories/DetalleDiagnosticoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DetalleDiagnosticoRepository extends JpaRepository<DetalleDiagnostico, Long> {
+public interface DetalleDiagnosticoRepository extends JpaRepository<DetalleDiagnostico, Integer> {
     public List<DetalleDiagnostico> findByDetalleCita(DetalleCita detalleCita);
     public List<DetalleDiagnostico> findByCodigoDiagnostico(Diagnostico diagnostico);
 }
